feat: Update serializer types from integer to float (#93297)
This PR fixes an issue where types coming from codecov were getting
flattened from floats to integers due to the
TestResultAggregatesSerializer defining the types incorrectly.


<!--

  Sentry employees and contractors can delete or ignore the following.

-->

### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.

+++ src/sentry/codecov/endpoints/TestResultsAggregates/serializers.py
@@ -11,16 +11,16 @@ class TestResultAggregatesSerializer(serializers.Serializer):
     Serializer for test results aggregates response
     """
 
-    totalDuration = serializers.IntegerField()
-    totalDurationPercentChange = serializers.IntegerField()
-    slowestTestsDuration = serializers.IntegerField()
-    slowestTestsDurationPercentChange = serializers.IntegerField()
+    totalDuration = serializers.FloatField()
+    totalDurationPercentChange = serializers.FloatField()
+    slowestTestsDuration = serializers.FloatField()
+    slowestTestsDurationPercentChange = serializers.FloatField()
     totalSlowTests = serializers.IntegerField()
-    totalSlowTestsPercentChange = serializers.IntegerField()
+    totalSlowTestsPercentChange = serializers.FloatField()
     totalFails = serializers.IntegerField()
-    totalFailsPercentChange = serializers.IntegerField()
+    totalFailsPercentChange = serializers.FloatField()
     totalSkips = serializers.IntegerField()
-    totalSkipsPercentChange = serializers.IntegerField()
+    totalSkipsPercentChange = serializers.FloatField()
 
     def to_representation(self, graphql_response):
         """
